feat(like): add repository methods and custom exceptions

+++ src/like/infrastructure/typeorm/repository/like.ts
@@ -2,6 +2,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { LikeRepository } from 'src/like/domain/repositories/like.repository';
 import { LikeModel } from '../models/like';
 import { Repository } from 'typeorm';
+import { NotFoundLikeException, RepeatLikeException } from 'src/like/domain/exceptions/like';
 
 export class TypeOrmLikeRepository implements LikeRepository {
   constructor(
@@ -10,14 +11,23 @@ export class TypeOrmLikeRepository implements LikeRepository {
   ) {}
 
   async addLike(userId: string, projectId: string): Promise<void> {
-    const like = new LikeModel();
-    like.userId = userId;
-    like.projectId = projectId;
-
+    const existing = await this.likeRepository.findOne({
+      where: { userId, projectId },
+    });
+    if (existing) {
+      throw new RepeatLikeException();
+    }
+    const like = this.likeRepository.create({ userId, projectId });
     await this.likeRepository.save(like);
   }
 
   async removeLike(userId: string, projectId: string): Promise<void> {
+    const existing = await this.likeRepository.findOne({
+      where: { userId, projectId },
+    });
+    if (!existing) {
+      throw new NotFoundLikeException();
+    }
     await this.likeRepository.delete({ userId, projectId });
   }
